Convert :setup to SetupCommand
Lift the :setup scaffolding handler into a standalone IFalloutCommand type that
takes IConsolePrompts via the constructor (first command to use injected prompts).
The scaffolding helpers (WriteBuildScripts/WriteConfigurationFile/GetTemplate and
the test-exercised UpdateSolutionFileContent) stay on Program as internal residual
until the #392 collapse PR. A transitional internal Setup shim remains on Program
so the still-legacy cake handler keeps compiling; it is removed when cake converts.
